examples/federated_learning/pyproject.toml
[project]
name = "aiq-federated-learning"
version = "0.1.0"
description = "Federated Learning Example for AIQToolkit"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "cryptography>=3.4.8",
    "websockets>=11.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.3.0"
]